various nitpicks
