+++ execution_engine/Cargo.toml
@@ -54,7 +54,6 @@ wasmi = "0.8.0"
 assert_matches = "1.3.0"
 criterion = "0.3.5"
 proptest = "1.0.0"
-proptest-attr-macro = "1.0.0"
 tempfile = "3.1.0"
 
 [features]
