+++ execution_engine/src/shared/newtypes/macros.rs
@@ -99,6 +99,7 @@ macro_rules! make_array_newtype {
             }
         }
 
+        #[allow(unused_qualifications)]
         impl bytesrepr::ToBytes for $name {
             fn to_bytes(&self) -> Result<Vec<u8>, bytesrepr::Error> {
                 self.0.to_bytes()
@@ -109,6 +110,7 @@ macro_rules! make_array_newtype {
             }
         }
 
+        #[allow(unused_qualifications)]
         impl bytesrepr::FromBytes for $name {
             fn from_bytes(bytes: &[u8]) -> Result<(Self, &[u8]), bytesrepr::Error> {
                 let (dat, rem) = <[$ty; $len]>::from_bytes(bytes)?;
