+++ node/src/components/consensus/consensus_protocol.rs
@@ -9,12 +9,13 @@ use anyhow::Error;
 use datasize::DataSize;
 use serde::{Deserialize, Serialize};
 
+use casper_hashing::Digest;
+use casper_types::bytesrepr::ToBytes;
+
 use crate::{
     components::consensus::{traits::Context, ActionId, TimerId},
     types::{TimeDiff, Timestamp},
 };
-use casper_hashing::Digest;
-use casper_types::bytesrepr::ToBytes;
 
 /// Information about the context in which a new block is created.
 #[derive(Clone, DataSize, Eq, PartialEq, Debug, Ord, PartialOrd, Hash)]
