+++ node/src/components/rpc_server/rpcs/chain/era_summary.rs
@@ -2,12 +2,11 @@ use once_cell::sync::Lazy;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
+use casper_hashing::Digest;
 use casper_types::{
     system::auction::{EraInfo, SeigniorageAllocation},
     AsymmetricType, EraId, PublicKey, U512,
 };
-
-use casper_hashing::Digest;
 
 use crate::{
     rpcs::{common::MERKLE_PROOF, docs::DocExample},
