+++ node/src/types/item.rs
@@ -7,10 +7,11 @@ use derive_more::Display;
 use serde::{de::DeserializeOwned, Serialize};
 use serde_repr::{Deserialize_repr, Serialize_repr};
 
-use crate::types::{BlockHash, BlockHeader, BlockHeaderWithMetadata};
 use casper_execution_engine::storage::trie::Trie;
 use casper_hashing::Digest;
 use casper_types::{bytesrepr::ToBytes, Key, StoredValue};
+
+use crate::types::{BlockHash, BlockHeader, BlockHeaderWithMetadata};
 
 /// An identifier for a specific type implementing the `Item` trait.  Each different implementing
 /// type should have a unique `Tag` variant.
